Turn DocumentFormat values into properties
Initializers on static fields are executed when the type is loaded, and some of the code pages being referenced don't exist off of Windows.

Changing them to properties causes them to load lazily. It also results in them not being cached, but perf is not a big concern for us.